refactor: 카드 생성
